Amend endpoints for enquires and restrict content to delete
The decision to restrict delete content is due to the abilty to
iterate.  This is definate functionality needed for the future,
but for now its better to keep this away from the user.

The problems with the endpoints included url encoding. Thishas been
addressed by send the email within the body of a delete.

Changes
- amend regression tests file
- add enquiries functionality in shared crud delete
- amend tests that broke
- rework handlers